+++ bean-searcher/src/main/java/cn/zhxu/bs/BeanSearcher.java
@@ -1,5 +1,7 @@
 package cn.zhxu.bs;
 
+import cn.zhxu.bs.util.FieldFns;
+
 import java.util.List;
 import java.util.Map;
 
@@ -12,6 +14,14 @@ import java.util.Map;
 public interface BeanSearcher extends Searcher {
 
 	/**
+	 * 适合需要分页的查询，此方法省略了检索参数，您可以从参数过滤器中注入它们
+	 * @param <T> bean 类型
+	 * @param beanClass 要检索的 bean 类型
+	 * @return 总条数，Bean 数据列表
+	 * */
+//	<T> SearchResult<T> search(Class<T> beanClass);
+
+	/**
 	 * 适合需要分页的查询
 	 * @param <T> bean 类型
 	 * @param beanClass 要检索的 bean 类型
@@ -19,6 +29,35 @@ public interface BeanSearcher extends Searcher {
 	 * @return 总条数，Bean 数据列表
 	 * */
 	<T> SearchResult<T> search(Class<T> beanClass, Map<String, Object> paraMap);
+
+	/**
+	 * 适合需要分页的查询，此方法省略了检索参数，您可以从参数过滤器中注入它们
+	 * @param <T> bean 类型
+	 * @param beanClass 要检索的 bean 类型
+	 * @param summaryFields 统计字段
+	 * @return 总条数，Bean 数据列表
+	 * */
+//	<T> SearchResult<T> search(Class<T> beanClass, String[] summaryFields);
+
+	/**
+	 * 适合需要分页的查询
+	 * @param <T> bean 类型
+	 * @param beanClass 要检索的 bean 类型
+	 * @param paraMap 检索参数
+	 * @param summaryField 统计字段
+	 * @return 总条数，Bean 数据列表
+	 * */
+	<T> SearchResult<T> search(Class<T> beanClass, Map<String, Object> paraMap, String summaryField);
+
+	/**
+	 * 适合需要分页的查询
+	 * @param <T> bean 类型
+	 * @param beanClass 要检索的 bean 类型
+	 * @param paraMap 检索参数
+	 * @param summaryField 统计字段
+	 * @return 总条数，Bean 数据列表
+	 * */
+	<T> SearchResult<T> search(Class<T> beanClass, Map<String, Object> paraMap, FieldFns.FieldFn<T, ?> summaryField);
 
 	/**
 	 * 适合需要分页的查询