+++ bean-searcher/src/main/java/cn/zhxu/bs/implement/DefaultBeanSearcher.java
@@ -4,6 +4,7 @@ import cn.zhxu.bs.*;
 import cn.zhxu.bs.bean.BeanAware;
 import cn.zhxu.bs.bean.ParamAware;
 import cn.zhxu.bs.param.FetchType;
+import cn.zhxu.bs.util.FieldFns;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -33,8 +34,27 @@ public class DefaultBeanSearcher extends AbstractSearcher implements BeanSearche
 	}
 
 	@Override
+	public <T> SearchResult<T> search(Class<T> beanClass, Map<String, Object> paraMap, String summaryField) {
+		if (summaryField != null) {
+			return search(beanClass, paraMap, new String[] { summaryField });
+		}
+		return search(beanClass, paraMap);
+	}
+
+	@Override
+	public <T> SearchResult<T> search(Class<T> beanClass, Map<String, Object> paraMap, FieldFns.FieldFn<T, ?> summaryField) {
+		if (summaryField != null) {
+			return search(beanClass, paraMap, FieldFns.name(summaryField));
+		}
+		return search(beanClass, paraMap);
+	}
+
+	@Override
 	public <T> SearchResult<T> search(Class<T> beanClass, Map<String, Object> paraMap, String[] summaryFields) {
-		return search(beanClass, paraMap, new FetchType(FetchType.DEFAULT, summaryFields));
+		if (summaryFields != null) {
+			return search(beanClass, paraMap, new FetchType(FetchType.DEFAULT, summaryFields));
+		}
+		return search(beanClass, paraMap);
 	}
 
 	@Override